% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_IO.R
\name{add_suffix_to_filename}
\alias{add_suffix_to_filename}
\title{A helper function that takes a file path and add a suffix to it}
\usage{
add_suffix_to_filename(file_path, suffix)
}
\arguments{
\item{file_path}{Character, file path to be transformed with the suffix}

\item{suffix}{Character, the suffix to be added to the file path}
}
\description{
A helper function that takes a file path and add a suffix to it
}
