% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_core.R
\name{area_function}
\alias{area_function}
\title{Calculate parameters of a given peak}
\usage{
area_function(data, xmin, xmax, alpha_beta, config)
}
\arguments{
\item{data}{data frame with the XRD pattern}

\item{xmin, xmax}{minimum and maximum values of where the desired Bragg peak is.
Used to isolate the desired Bragg peak from a others.}

\item{alpha_beta}{flag to control the type of radiation source
=2 if monochromatic, =3 if Kalpha doublet}

\item{config}{List, configuration file}
}
\value{
a list with:
"pattern_lim" a data frame with the 2theta reduced and background
subtracted XRD pattern and column expanded to include the values of
the spline fit function over data, the normalized pattern, and
the normalized spline fitted data
"integ_result" a list with the integration results of the Bragg peak (through spline fit)
"area_total" a double with the result of calculated area of the Bragg peak
"integ_result_norm" a list with the integration results of the normalized Bragg peak (through spline fit)
"twoTheta_max" a double with the 2theta position of the Bragg peak maximum
"beta" a double with the integral breadth of the normalized Bragg peak
"max_peak_rest" a double with the max intensity of the Bragg peak after background subtraction
"beta_unorm" a double with the integral breadth of the Bragg peak before normalization.
}
\description{
Calculate parameters of a given peak
}
