% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_correction.R
\name{bkg_subtract_sym_function}
\alias{bkg_subtract_sym_function}
\title{Subtract background from data}
\usage{
bkg_subtract_sym_function(data, max_position, dist_from_center = 0.7)
}
\arguments{
\item{data}{XRD pattern, already cleaned (2theta range filtered to be centered at the maximum of Bragg peak)}

\item{max_position}{2theta value where the maximum of Bragg peak is}

\item{dist_from_center}{Distance to the 2theta position of the Bragg peak maximum
to exclude for the background fitting. The idea is that all 2theta in the range:
max_position - dist_from_center <= 2theta <= max_position + dist_from_center
are well inside the Bragg peak, and those outiside this range could be safely
considered at the background.}
}
\value{
A list with an expanded "data" data frame with two additional columns:
"bkg" with the background values at each 2theta and "int_rest" with the
background subtracted intensity values, therefore columns (2theta, int_rest) are the
background subtracted XRD pattern
}
\description{
Subtract background from data
}
