% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_utils.R
\name{create_config_template}
\alias{create_config_template}
\title{Create a new configuration template}
\usage{
create_config_template(output_file = "config/my_analysis.yml")
}
\arguments{
\item{output_file}{file name and path of the new configuration file}
}
\description{
Create a new configuration template
}
