% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_correction.R
\name{.create_background_model}
\alias{.create_background_model}
\title{Create linear background function for subtraction}
\usage{
.create_background_model(x, a, b)
}
\arguments{
\item{x}{X-value to calculate function at}

\item{a, b}{coefficients of linear function y = a + b * x}
}
\value{
Background value at x
}
\description{
Create linear background function for subtraction
}
\keyword{internal}
