% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_IO.R
\name{export_sequential_csv}
\alias{export_sequential_csv}
\title{Exporting results to csv file (sequential processing version)}
\usage{
export_sequential_csv(data_list, output_directory, sample_name, measurement)
}
\arguments{
\item{data_list}{List or tibble, data to export to csv file}

\item{output_directory}{Character, path to destination directory}

\item{sample_name}{Character, name of the measured sample (to construct file name)}

\item{measurement}{Character, measurement identifier (to construct file name)}
}
\description{
Exporting results to csv file (sequential processing version)
}
