% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_core.R
\name{extract_results_function}
\alias{extract_results_function}
\title{Results aggregation}
\usage{
extract_results_function(nested_list)
}
\arguments{
\item{nested_list}{A nested list with the results from several XRD diffraction patterns
analysis. Main list goes along the different patterns, inside of each pattern
list is another list with the pattern-specific results}
}
\value{
A tibble with its columns being relevant peak parameters and each row corresponding
to a XRD pattern.
}
\description{
Results aggregation
}
