% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_IO.R
\name{import_all_data}
\alias{import_all_data}
\title{Main import function}
\usage{
import_all_data(config)
}
\arguments{
\item{config}{List, containing data input configuration (config$data_input$...)}
}
\value{
Nested list of imported data by subdirectory (e.g. by "up", "down")
}
\description{
Main import function
}
