% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_IO.R
\name{import_xrdml}
\alias{import_xrdml}
\title{Load a difraction pattern from a panalytical file (.xrdml) measured in a non-ambient setup,
extract the proper diffraction pattern (2theta, intensity) and store it in a data frame,
extract the temperature and store it an double. Return a list with (pattern, temperature).}
\usage{
import_xrdml(file_name)
}
\arguments{
\item{file_name}{Complete path and file name of the .xrdml file wanted to be imported}
}
\value{
A list with a data frame and a double.
Data frame is the XRD pattern (2 columns: 2theta, intensity)
Double is the temperature where pattern was acquired
}
\description{
Load a difraction pattern from a panalytical file (.xrdml) measured in a non-ambient setup,
extract the proper diffraction pattern (2theta, intensity) and store it in a data frame,
extract the temperature and store it an double. Return a list with (pattern, temperature).
}
\examples{
\dontrun{
difractogram_data <- import_xrdml("/data/sample/xrd_pattern.xrdml")
}
}
