% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_correction.R
\name{interp_points}
\alias{interp_points}
\title{Linear interpolation between two points (possible not in use!!!)}
\usage{
interp_points(x1, y1, x2, y2, x)
}
\arguments{
\item{x1, y1}{Coordinates of first point}

\item{x2, y2}{Coordinates of second point}

\item{x}{X-value to interpolate at}
}
\value{
Interpolated y-value
}
\description{
Linear interpolation between two points (possible not in use!!!)
}
