% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_workflow.R
\name{run_complete_analysis}
\alias{run_complete_analysis}
\title{This is the main analysis function (orchestration) that is called by the runner. It takes a configuration file and an environment to proceed with the complete analysis pipeline.}
\usage{
run_complete_analysis(
  config_file = "analysis_config.yml",
  env = NULL,
  override_list = NULL
)
}
\arguments{
\item{config_file}{Character, name of the configuration file to be used}

\item{env}{Character, environment to be used (production == NULL, development, testing)}

\item{override_list}{List, a list that overrides the config_file in case to be needed}
}
\description{
This is the main analysis function (orchestration) that is called by the runner. It takes a configuration file and an environment to proceed with the complete analysis pipeline.
}
