% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_utils.R
\name{save_config_with_results}
\alias{save_config_with_results}
\title{Save configuration with results for reproducibility}
\usage{
save_config_with_results(config, output_dir)
}
\arguments{
\item{config}{configuration file to be saved}

\item{output_dir}{path where configuration file is going to be saved to}
}
\description{
Save configuration with results for reproducibility
}
