% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_IO.R
\name{single_plot_function}
\alias{single_plot_function}
\title{Plotting results using ggplot2 library. It is expected to work with 2 csv files ("up" and "down" directions) or with a list of 2 data frames ("up" and "down" directions)}
\usage{
single_plot_function(filepointer, method = c("csv", "list"), saving_dir)
}
\arguments{
\item{filepointer}{path where the csv files intended for plot are}

\item{method}{flag to control what is the input, csv files (method == "csv") or a list of data frames (method == "list").}

\item{saving_dir}{path to output directory where graph file is saved}
}
\value{
return a ggplot2 object with the plot, additionally it saves it in output directory as a png file
}
\description{
Plotting results using ggplot2 library. It is expected to work with 2 csv files ("up" and "down" directions) or with a list of 2 data frames ("up" and "down" directions)
}
\examples{
\dontrun{
  single_plot_function(path_to_csv_files, method = "csv", path_to_output_dir)
  single_plot_function(list_name, method = "list", path_to_output_dir)
}

}
